+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 
+- Add AMW deduplicating flag, SentToAMW, for AKS Attach
+  ([#41410](https://github.com/Azure/azure-sdk-for-python/pull/41410))
+
 ### Breaking Changes
 
 ### Bugs Fixed